refactor: ポータル機能を router へ分離
